refactor(CLI): Split into more digestable files
